requirements.txt
asgiref==3.8.1
Django==5.1.2
django-resized==1.0.3
pillow==10.4.0
sqlparse==0.5.1
